#49 float64からfloat32に変更

+++ Backend/api/internal/db/spacetime/table.go
@@ -3,8 +3,8 @@ package spacetime
 type SpaceTime struct {
 	Id         int     `db:"id"`
 	Time       string  `db:"time"`
-	Latitude   float64 `db:"latitude"`
-	Longtitude float64 `db:"longtitude"`
-	Altitude   float64 `db:"altitude"`
+	Latitude   float32 `db:"latitude"`
+	Longtitude float32 `db:"longtitude"`
+	Altitude   float32 `db:"altitude"`
 	ObjId      int     `db:"obj_id"`
 }
